+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- Add `<style_sheet_hash>`, `<style_sheet_hash_short>`, `<name_hash>`, and `<name_hash_short>` placeholders into `package.metadata.turf.class_names.template` configuration option
+
 # 0.9.3
 
 - Updated lightningcss